Aggiungi DTO per le tariffe medici e il trend mensile; modifica Medico e Prenotazione per gestire nuovi campi e relazioni; implementa il metodo getAll in MedicoService e il relativo endpoint in MedicoController.

+++ H-Gate-api/src/main/java/com/development/spring/hGate/H_Gate/dtos/prenotazioni/SlotDisponibiliDTO.java
@@ -13,7 +13,7 @@ import java.util.List;
 @AllArgsConstructor
 public class SlotDisponibiliDTO {
 
-    private Long medicoId;
+    private Integer medicoId;
     private String medicoNome;
     private String data;
     private List<SlotDisponibileDTO> slots;
